+++ NEWS.md
@@ -60,6 +60,10 @@ Standard library changes
 
 #### Profile
 
+#### Random
+
+* `shuffle` now may take an argument of `NTuple` value ([#56906]).
+
 #### REPL
 
 * The display of `AbstractChar`s in the main REPL mode now includes LaTeX input information like what is shown in help mode ([#58181]).